Refactoring: tagging unit tests with @SpringBootUnitTest that includes other tags like @SpringBootTest, @Tag nad @ActiveProfile.
