+++ fhir-server/src/test/java/au/csiro/pathling/update/BatchProviderTest.java
@@ -28,6 +28,7 @@ import static org.mockito.Mockito.verifyNoInteractions;
 import au.csiro.pathling.config.ServerConfiguration;
 import au.csiro.pathling.errors.InvalidUserInputError;
 import au.csiro.pathling.io.Database;
+import au.csiro.pathling.test.SpringBootUnitTest;
 import ca.uhn.fhir.parser.IParser;
 import java.util.List;
 import javax.annotation.Nonnull;
@@ -35,15 +36,12 @@ import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.Enumerations.ResourceType;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatcher;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
-@SpringBootTest
-@Tag("UnitTest")
+@SpringBootUnitTest
 class BatchProviderTest {
 
   @Autowired
